+++ tests/download_attachment.php
@@ -1,0 +1,3 @@
+<?php
+
+header('Location: ' . $_GET['download_url'], true, 303);